Remove debugging settings. More tests.
