+++ src/test/resources/simplelogger.properties
@@ -1,3 +1,4 @@
 org.slf4j.simpleLogger.showDateTime=true
 org.slf4j.simpleLogger.defaultLogLevel=info
 org.slf4j.simpleLogger.log.com.github.kagkarlsson.scheduler=info
+org.slf4j.simpleLogger.log.com.github.kagkarlsson.scheduler.DueExecutionsBatch=info
